+++ src/index.html
@@ -3,13 +3,13 @@
 	<head>
 		<title>Restaurant Page</title>
 		<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-BmbxuPwQa2lc/FVzBcNJ7UAyJxM6wuqIj61tLrc4wSX0szH/Ev+nYRRuWlolflfl" crossorigin="anonymous">
-		<link rel="stylesheet" href="./index.css">
+		<link rel="stylesheet" href="./main.css">
 	</head>
 	<body>
 
 		<div id="content">
 
 		</div>
-		<script src="./app.js"></script>
+		<script type="module" src="./app.js"></script>
 	</body>
 </html>
